# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SFA - Air Charter Service API",
        "description": "API para o sistema de fretamento de voos SFA.",
        "contact": {
            "email": "admin@sfa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/flights": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get list of flights",
                "description": "Returns list of flights",
                "operationId": "getFlightsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FlightResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Store new flight",
                "description": "Returns flight data",
                "operationId": "storeFlight",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFlightRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlightResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flights/{id}": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get flight information",
                "description": "Returns flight data",
                "operationId": "getFlightById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlightResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Flights"
                ],
                "summary": "Update existing flight",
                "description": "Returns updated flight data",
                "operationId": "updateFlight",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFlightRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlightResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Flights"
                ],
                "summary": "Delete existing flight",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteFlight",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "FlightResource": {
                "title": "Flight Resource",
                "description": "Flight resource model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "SFA123"
                    },
                    "origin": {
                        "type": "string",
                        "example": "New York (JFK)"
                    },
                    "destination": {
                        "type": "string",
                        "example": "Los Angeles (LAX)"
                    },
                    "departure_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-01T10:00:00Z"
                    },
                    "arrival_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-01T13:00:00Z"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 450.75
                    },
                    "aircraft": {
                        "type": "string",
                        "example": "Boeing 737"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "scheduled",
                            "cancelled",
                            "completed"
                        ],
                        "example": "scheduled"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "StoreFlightRequest": {
                "title": "Store Flight Request",
                "description": "Request body for creating or updating a flight",
                "required": [
                    "code",
                    "origin",
                    "destination",
                    "departure_time",
                    "arrival_time",
                    "price",
                    "aircraft",
                    "status"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "SFA124"
                    },
                    "origin": {
                        "type": "string",
                        "example": "Chicago (ORD)"
                    },
                    "destination": {
                        "type": "string",
                        "example": "Miami (MIA)"
                    },
                    "departure_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-15T14:00:00Z"
                    },
                    "arrival_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-15T17:30:00Z"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 320.5
                    },
                    "aircraft": {
                        "type": "string",
                        "example": "Airbus A320"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "scheduled",
                            "cancelled",
                            "completed"
                        ],
                        "example": "scheduled"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}